% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutledge-package.R
\docType{package}
\name{rutledge-package}
\alias{rutledge-package}
\title{rutledge: Real-Time PCR Data Sets by Rutledge et al. (2004)}
\description{
Real-time quantitative polymerase chain reaction (qPCR) data by Rutledge et al. (2004) \doi{10.1093/nar/gnh177} in tidy format. The data comprises a six-point, ten-fold dilution series, repeated in five independent runs, for two different amplicons. In each run, each standard concentration is replicated four times. For the original raw data file see the Supplementary Data section: \url{https://academic.oup.com/nar/article/32/22/e178/2375678#supplementary-data}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/rutledge}
  \item \url{https://rmagno.eu/rutledge/}
  \item Report bugs at \url{https://github.com/ramiromagno/rutledge/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
