% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutledge-package.R
\docType{package}
\name{rutledge-package}
\alias{rutledge-package}
\alias{_PACKAGE}
\title{rutledge: Real-Time PCR Data Sets by Rutledge et al. (2004)}
\description{
Real-time quantitative polymerase chain reaction (qPCR) data by Rutledge et al. (2004) \doi{10.1093/nar/gnh177} in tidy format. The data comprises a six-point, ten-fold dilution series, repeated in five independent runs, for two different amplicons. In each run, each standard concentration is replicated four times. Original raw data file: \url{https://oup.silverchair-cdn.com/oup/backfile/Content_public/Journal/nar/32/22/10.1093_nar_gnh177/1/gnh177_Supplementary_Data.zip?Expires=1680370010&Signature=xxk4VxX0C4yr3UWzxgh7ieCt7QQmpFMRauvsVEwtGXYGCcyQY7uhNCE-M8zx9kpyDPoS8NR7fjBuMx2Xz2ANFwF1VqnjQ4AfO37klnJ3CHRIJ7bj01n2mycHDnvJ3XawHdWT8TqJxTxVC9CpYEkH2xGeJBnnpcnXLbc94A8KB8FCtg2WR3O~ULkaOQQ8uJAiVdJhnBHH~XfBRkfoKHSuyJgX7n7M2~gwXnZH9n3vUyo~CHrpIax7Hi0xUSCBbQM571hxA7JIHkhZ0HBm2aXFuAru2yJ~o8jMEnnguOJg8T7mGqTDzUBtW0hJhmQDksdJoyeAFzU84QRUIZj9q3-tXg__&Key-Pair-Id=APKAIE5G5CRDK6RD3PGA}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/rutledge}
  \item Report bugs at \url{https://github.com/ramiromagno/rutledge/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
